code clean fixes

+++ src/main/java/functions/FileFunctions.java
@@ -39,11 +39,8 @@ public class FileFunctions extends BaseFunction {
 
   public void captureScreenshotOfElement(String fileName, WebElement element) {
     log.debug("Capture image of element {" + getElementInfo(element) + "}");
-
     File target = new File(pathOutputFolder.toAbsolutePath().toString() + "/" + fileName + ".png");
-    if (target.exists())
-      target.delete();
-
+    if (target.exists()) target.delete();
     File screen = ((TakesScreenshot) driver.getDriver()).getScreenshotAs(OutputType.FILE);
 
     Point p = element.getLocation();
@@ -55,7 +52,6 @@ public class FileFunctions extends BaseFunction {
       BufferedImage dest = img.getSubimage(p.getX(), p.getY(), width, height);
       ImageIO.write(dest, "png", screen);
       Files.copy(screen.toPath(), target.toPath());
-
       log.debug("File copied to {" + pathOutputFolder.toAbsolutePath().toString() + "\\" + fileName + ".png}");
     } catch (IOException e) {
       e.printStackTrace();
@@ -68,7 +64,6 @@ public class FileFunctions extends BaseFunction {
 
   public void saveTextToFile(String textValue, String fileName, boolean append) {
     File target = new File(pathOutputFolder.toAbsolutePath().toString() + "/" + fileName + ".txt");
-
     try (FileWriter fw = new FileWriter(target, append);
          BufferedWriter bw = new BufferedWriter(fw);
          PrintWriter out = new PrintWriter(bw)) {
@@ -95,7 +90,6 @@ public class FileFunctions extends BaseFunction {
         sb.append(ls);
       }
       reader.close();
-
       result = sb.toString();
     } catch (FileNotFoundException e) {
       log.error(e.getMessage());
@@ -103,7 +97,6 @@ public class FileFunctions extends BaseFunction {
     } catch (IOException e) {
       e.printStackTrace();
     }
-
     return result;
   }
 
@@ -127,7 +120,6 @@ public class FileFunctions extends BaseFunction {
         e.printStackTrace();
       }
     }
-
     log.debug("Downloaded {" + counter + " images}");
   }
